test(streaming): isolate resident done regression pin
